4.23 tehty

+++ controllers/blogs.js
@@ -3,13 +3,13 @@ const Blog = require('../models/blog');
 const User = require('../models/user');
 const jwt = require('jsonwebtoken');
 
-const getTokenFrom = (request) => {
-    const authorization = request.get('authorization');
-    if (authorization && authorization.startsWith('Bearer ')) {
-        return authorization.replace('Bearer ', '');
-    }
-    return null;
-};
+// const getTokenFrom = (request) => {
+//     const authorization = request.get('authorization');
+//     if (authorization && authorization.startsWith('Bearer ')) {
+//         return authorization.replace('Bearer ', '');
+//     }
+//     return null;
+// };
 
 blogsRouter.get('', async (request, response) => {
     const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 });
@@ -17,8 +17,20 @@ blogsRouter.get('', async (request, response) => {
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndDelete(request.params.id);
-    response.status(204).end();
+    const blog = await Blog.findById(request.params.id);
+    const user = request.user;
+    const decodedToken = jwt.verify(request.token, process.env.SECRET);
+    if (!decodedToken || !decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' });
+    }
+    if (blog.user.toString() === user.id.toString()) {
+        await Blog.findByIdAndDelete(request.params.id);
+        response.status(204).end();
+    } else {
+        response.status(401).send({ error: 'unauthorized' });
+    }
+    // await Blog.findByIdAndDelete(request.params.id);
+    // response.status(204).end();
 });
 
 blogsRouter.put('/:id', async (request, response) => {
@@ -35,11 +47,12 @@ blogsRouter.put('/:id', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
     const body = request.body;
-    const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET);
-    if (!decodedToken || !decodedToken.id) {
-        return response.status(401).json({ error: 'token missing or invalid' });
-    }
-    const user = await User.findById(decodedToken.id);
+    // const decodedToken = jwt.verify(request.token, process.env.SECRET);
+    // if (!decodedToken || !decodedToken.id) {
+    //     return response.status(401).json({ error: 'token missing or invalid' });
+    // }
+    // const user = await User.findById(decodedToken.id);
+    const user = request.user;
     if (!user) {
         return response.status(400).send({ error: 'user not found' });
     }
